+++ backend/math/src/functions/exponential.rs
@@ -1,0 +1,6 @@
+use crate::Number;
+
+pub struct ExponentialFunction {
+    pub c: Number,
+    pub a: Number,
+}