+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - f(x) graph problems
 - `dot_with_lines` for graphs
 - Graph legend support
+- Support for quadratic functions
 
 ## [0.12.5] - 2026-08-07